[CALCITE-3435] Enable decimal modulus operation to allow numeric with non-zero scale (DonnyZone)
Decimal modulus return type inference:
e1[s1, p1] % e2[s2, p2]

s = max(s1, s2)
p = min(p1-s1, p2-s2) + max(s1, s2)

Fix up(by Danny):
* Move the RelDataTypeSystemImpl#deriveDecimalModType impl to its parent interface as a
default
* Fix some typos

close apache/calcite#1563

+++ core/src/main/java/org/apache/calcite/runtime/SqlFunctions.java
@@ -1156,8 +1156,8 @@ public class SqlFunctions {
   }
 
   // temporary
-  public static int mod(int b0, BigDecimal b1) {
-    return mod(b0, b1.intValue());
+  public static BigDecimal mod(int b0, BigDecimal b1) {
+    return mod(BigDecimal.valueOf(b0), b1);
   }
 
   public static BigDecimal mod(BigDecimal b0, BigDecimal b1) {
